requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
Pillow>=9.5.0
pdf2image>=1.17.0
pytesseract>=0.3.10
pandas>=1.5.0